Copy more code over from DHT
